pyproject.toml
[tool.poetry]
name = "response-agent"
version = "0.1.0"
description = ""
authors = ["Marcio <marciofnogueira@gmail.com>"]
readme = "README.md"
package-mode = false  # Desativar o modo de pacote

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.2.5"
pandas = "^2.2.2"
pypdf2 = "^3.0.1"
streamlit = "^1.35.0"
openai = "^0.27.0"
pymongo = "^4.7.3"

[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
